--- a/modules/core/src/lib/model.js
+++ b/modules/core/src/lib/model.js
@@ -238,12 +238,19 @@ export default class Model {
 
     const logPriority = this._logDrawCallStart(2);
 
-    const drawParams = this.vertexArray.getDrawParams(this.props);
-    if (drawParams.isInstanced && !this.isInstanced) {
+    const drawParams = this.vertexArray.getDrawParams();
+    const {
+      isIndexed = drawParams.isIndexed,
+      indexType = drawParams.indexType,
+      indexOffset = drawParams.indexOffset,
+      vertexArrayInstanced = drawParams.isInstanced
+    } = this.props;
+
+
+    if (vertexArrayInstanced && !this.isInstanced) {
       log.warn('Found instanced attributes on non-instanced model', this.id)();
     }
 
-    const {isIndexed, indexType, indexOffset} = drawParams;
     const {isInstanced, instanceCount} = this;
 
     const noop = () => {};