--- a/modules/webgl/src/classes/program.js
+++ b/modules/webgl/src/classes/program.js
@@ -6,6 +6,7 @@ import Framebuffer from './framebuffer';
 import {parseUniformName, getUniformSetter} from './uniforms';
 import {VertexShader, FragmentShader} from './shader';
 import ProgramConfiguration from './program-configuration';
+import {copyUniform, checkUniformValues} from './uniforms';
 
 import {withParameters} from '../context';
 import {assertWebGL2Context, isWebGL2, getKey} from '../webgl-utils';
@@ -53,9 +54,10 @@ export default class Program extends Resource {
   }
 
   initialize(props = {}) {
-    const {hash, vs, fs, varyings, bufferMode = GL_SEPARATE_ATTRIBS} = props;
+    const {hash, vs, fs, varyings, bufferMode = GL_SEPARATE_ATTRIBS, debug = false} = props;
 
     this.hash = hash || ''; // Used by ProgramManager
+    this.debug = debug;
 
     // Create shaders if needed
     this.vs =
@@ -184,34 +186,46 @@ export default class Program extends Resource {
     return true;
   }
 
-  setUniforms(uniforms = {}, _onChangeCallback = () => {}) {
-    // Simple change detection - if all uniforms are unchanged, do nothing
-    // let somethingChanged = false;
-    // const changedUniforms = {};
-    // for (const key in uniforms) {
-    //   if (!areUniformsEqual(this.uniforms[key], uniforms[key])) {
-    //     somethingChanged = true;
-    //     changedUniforms[key] = uniforms[key];
-    //     this.uniforms[key] = getUniformCopy(uniforms[key]);
-    //   }
-    // }
+  setUniforms(uniforms = {}) {
+    if (this.debug) {
+      checkUniformValues(uniforms, this.id, this._uniformSetters);
+    }
 
-    // if (somethingChanged) {
-    //   _onChangeCallback();
-    //   checkUniformValues(changedUniforms, this.id, this._uniformSetters);
-    //   this._setUniforms(changedUniforms);
-    // }
+    this.gl.useProgram(this.handle);
 
-    this._setUniforms(uniforms);
-    Object.assign(this.uniforms, uniforms);
+    for (const uniformName in uniforms) {
+      const uniform = uniforms[uniformName];
+      const uniformSetter = this._uniformSetters[uniformName];
+
+      if (uniformSetter) {
+        let value = uniform;
+        if (value instanceof Framebuffer) {
+          value = value.texture;
+        }
+        if (value instanceof Texture) {
+          // eslint-disable-next-line max-depth
+          if (uniformSetter.textureIndex === undefined) {
+            uniformSetter.textureIndex = this._textureIndexCounter++;
+          }
+
+          // Bind texture to index
+          const texture = value;
+          const {textureIndex} = uniformSetter;
+
+          texture.bind(textureIndex);
+          value = textureIndex;
+        }
+
+        if (uniformSetter(value)) {
+          copyUniform(this.uniforms, uniformName, uniform);
+        }
+      }
+    }
 
     return this;
   }
 
   // PRIVATE METHODS
-
-  // stub for shader chache, should reset uniforms to default valiues
-  reset() {}
 
   // Checks if all texture-values uniforms are renderable (i.e. loaded)
   // Note: This is currently done before every draw call
@@ -259,43 +273,6 @@ export default class Program extends Resource {
         }
       }
     }
-  }
-
-  // Apply a set of uniform values to a program
-  // Only uniforms actually present in the linked program will be updated.
-  _setUniforms(uniforms) {
-    this.gl.useProgram(this.handle);
-
-    for (const uniformName in uniforms) {
-      let uniform = uniforms[uniformName];
-      const uniformSetter = this._uniformSetters[uniformName];
-
-      if (uniformSetter) {
-        if (uniform instanceof Framebuffer) {
-          uniform = uniform.texture;
-        }
-        if (uniform instanceof Texture) {
-          // eslint-disable-next-line max-depth
-          if (uniformSetter.textureIndex === undefined) {
-            uniformSetter.textureIndex = this._textureIndexCounter++;
-          }
-
-          // Bind texture to index
-          const texture = uniform;
-          const {textureIndex} = uniformSetter;
-
-          texture.bind(textureIndex);
-
-          // Set the uniform sampler to the texture index
-          uniformSetter(textureIndex);
-        } else {
-          // Just set the value
-          uniformSetter(uniform);
-        }
-      }
-    }
-
-    return this;
   }
 
   // RESOURCE METHODS