--- a/modules/webgl/src/classes/program.js
+++ b/modules/webgl/src/classes/program.js
@@ -70,6 +70,8 @@ export default class Program extends Resource {
     // uniforms
     this.uniforms = {};
 
+    this._texturesRenderable = true;
+
     // Setup varyings if supplied
     if (varyings && varyings.length > 0) {
       assertWebGL2Context(this.gl);
@@ -129,7 +131,7 @@ export default class Program extends Resource {
       this.setUniforms(uniforms || {});
     }
 
-    if (logPriority !== undefined) {
+    if (this.debug && logPriority !== undefined) {
       const fb = framebuffer ? framebuffer.id : 'default';
       const message =
         `mode=${getKey(this.gl, drawMode)} verts=${vertexCount} ` +
@@ -216,6 +218,10 @@ export default class Program extends Resource {
           texture.bind(textureIndex);
           value = textureIndex;
 
+          if (!texture.loaded) {
+            this._texturesRenderable = false;
+          }
+
           textureUpdate = this.uniforms[uniformName] !== uniform;
         }
 
@@ -235,7 +241,11 @@ export default class Program extends Resource {
   // Checks if all texture-values uniforms are renderable (i.e. loaded)
   // Note: This is currently done before every draw call
   _areTexturesRenderable() {
-    let texturesRenderable = true;
+    if (this._texturesRenderable) {
+      return true;
+    }
+
+    this._texturesRenderable = true;
 
     for (const uniformName in this.uniforms) {
       const uniformSetter = this._uniformSetters[uniformName];
@@ -251,12 +261,12 @@ export default class Program extends Resource {
         if (uniform instanceof Texture) {
           const texture = uniform;
           // Check that texture is loaded
-          texturesRenderable = texturesRenderable && texture.loaded;
+          this._texturesRenderable = this._texturesRenderable && texture.loaded;
         }
       }
     }
 
-    return texturesRenderable;
+    return this._texturesRenderable;
   }
 
   // Binds textures