A round of optimizations:
- Uniform setting/diffing: separate setter functions based on whether the uniform will be set as an array or not, so you don't have to do those checks on each update.
- Move logging/debug info in `Program` and `Model` behind debug flags.
- Track whether textures are loading in a property so the full check doesn't have to be done on every draw.
- Avoid allocating in `Model.draw` and `Program.draw`

Some API changes to support this:
- `VertexArray.getDrawParams` no longer takes overrides as an argument. Overriding can be done by the caller.
- After a uniform is set, it's size cannot be changed. Only case where this would occur is if application was setting both an entire uniform array and the first element using the same name.

Results on the stress test:

Metric  |Before|After|Improvement
--------|------|-----|-----------
CPU Time|32ms  |22ms |10ms (31%)
GPU Time|28ms  |24ms |4ms  (14%)